Revise Docs CG charter scope of work
At the first Docs CG meeting, I suggested that initial draft of the
charter did not capture the spirit of the work under discussion or the
relationship of the CG to other organizations and efforts. I'm proposing
a slimmed-down version of the CG's scope of work, to focus on
coordination, relationship building, and information sharing, rather
than on docs writing and maintenance per se.

I've also dropped the key deliverables sections, since I think it's
slightly premature to formalize what those deliverables might be, in
light of the revised scope of work.